handle the case of source being a dir in copy

+++ arcos_backend/filesystem/__init__.py
@@ -62,8 +62,12 @@ class Filesystem:
                     self._root.joinpath(destination))
 
     def copy(self, source: PathLike | str, destination: PathLike | str):
-        shutil.copy(self._root.joinpath(source),
-                    self._root.joinpath(destination))
+        if Path(source).is_file():
+            shutil.copy(self._root.joinpath(source),
+                        self._root.joinpath(destination))
+        else:
+            shutil.copytree(self._root.joinpath(source),
+                            self._root.joinpath(destination))
 
     def read(self, path: PathLike | str) -> bytes:
         return self._root.joinpath(path).read_bytes()
